Rename color_dicts.py to color.py.
