+++ gmplot/color.py
